Move bridges into bridge folder
